+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Changelog
 
 * Added `K.getValue()` function.
 * Upgraded to Kefir 3.6.1
+* Fix a bug where a bus would not signal its end when subscribing to
+  `bus.stream.onEnd()` _after_ calling `bus.end()`.
 * `K.combinePropertyObject()` now returns a proper property
 
 ## v0.5.0 2016-10-16